clinica-medica/requirements.txt
fastapi
sqlmodel
uvicorn
alembic
python-dotenv
